JSON Serialization & Custom Tests
Draw any image, save it to a named folder and they will be added to a custom test list. Currently, there is no way to use the test data, but that will come soon. As for JSON serialization, the neural network can be saved to a (large) data file (@50,000 lines) and then also loaded back in to continue where it left off. The network is saved automatically every epoch by default.